backend_django/requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
django-cors-headers>=4.3
gunicorn>=21.0
whitenoise>=6.6
pymongo>=4.6
